==> notebooks/05_aws.ipynb <==
import boto3
from pathlib import Path

# ---- AWS Configuration ----
BUCKET_NAME = "weather-forecast-data-mle"  
REGION = "eu-central-1" 

PROJECT_ROOT = Path.cwd().parent
PROCESSED_DIR = PROJECT_ROOT / "data" / "processed"
MODEL_DIR = PROJECT_ROOT / "models"

# Initialize S3 Client
s3 = boto3.client("s3", region_name=REGION)

def upload_file(local_path: Path, s3_prefix: str, mae: float = None):
    s3_key = f"{s3_prefix}/{local_path.name}"
    
    # Adding metadata allows you to see the model performance in the AWS Console
    extra_args = {}
    if mae:
        extra_args = {'Metadata': {'mae': str(mae)}}
    
    print(f"Uploading {local_path.name} → s3://{BUCKET_NAME}/{s3_key}")
    s3.upload_file(str(local_path), BUCKET_NAME, s3_key, ExtraArgs=extra_args)


upload_file(PROCESSED_DIR / "train_encoded.csv", "processed")
upload_file(PROCESSED_DIR / "eval_encoded.csv", "processed")
upload_file(PROCESSED_DIR / "holdout_encoded.csv", "processed")


upload_file(MODEL_DIR / "best_weather_xgb.pkl", "models")
upload_file(MODEL_DIR / "tuned_scaler.pkl", "models")

print("\nAll weather pipeline artifacts pushed to AWS S3.")
